+++ FosterRoster/Data/FosterRosterDbContext.cs
@@ -5,9 +5,11 @@ using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using Features.Account;
 using Features.Chores;
+using Features.Comments;
 using Features.Felines;
 using Features.Fosterers;
 using Features.Sources;
+using Features.Thumbnails;
 using Features.Weights;
 
 public class FosterRosterDbContext(DbContextOptions<FosterRosterDbContext> options)
