+++ FosterRoster/Features/Comments/CommentCard.razor
@@ -1,6 +1,3 @@
-@inject TimeProvider TimeProvider
-@inject ILogger<CommentCard> Logger
-
 @if (_model is null)
 {
     <RadzenCard Variant="Variant.Outlined" class="rz-py-1 rz-px-2">
@@ -33,8 +30,8 @@ else
 }
 
 @code {
-    private MarkupString? _html = null;
-    private CommentEditModel? _model = null;
+    private MarkupString? _html;
+    private CommentEditModel? _model;
 
     [EditorRequired] [Parameter] public Comment Comment { get; set; } = null!;
