+++ FosterRoster/Features/Comments/CommentRepository.cs
@@ -1,7 +1,7 @@
-namespace FosterRoster.Features.Felines;
+namespace FosterRoster.Features.Comments;
 
 public sealed class CommentRepository(
-    IDbContextFactory<Data.FosterRosterDbContext> contextFactory,
+    IDbContextFactory<Data.FosterRosterDbContext> dbContextFactory,
     TimeProvider timeProvider
 )
 {
@@ -12,14 +12,14 @@ public sealed class CommentRepository(
     /// <returns>A Result with Comment on Success, otherwise Result with Errors.</returns>
     public async Task<Result<Comment>> AddAsync(Comment comment)
     {
-        await using var context = await contextFactory.CreateDbContextAsync();
+        await using var context = await dbContextFactory.CreateDbContextAsync();
         // Workaround because of issues getting ValueGeneratedOnAdd() to work.
         comment.TimeStamp = timeProvider.GetUtcNow().UtcDateTime;
         var entry = await context.Comments.AddAsync(comment);
         await context.SaveChangesAsync();
         return Result.Ok(entry.Entity);
     }
-
+    
     /// <summary>
     ///     Removes an existing comment by its primary key.
     /// </summary>
@@ -27,7 +27,7 @@ public sealed class CommentRepository(
     /// <returns>A Result instance indicating success or failure.</returns>
     public async Task<Result> DeleteByKeyAsync(int commentId)
     {
-        await using var context = await contextFactory.CreateDbContextAsync();
+        await using var context = await dbContextFactory.CreateDbContextAsync();
         return await context
                 .Comments
                 .Where(c => c.Id == commentId)
@@ -47,7 +47,7 @@ public sealed class CommentRepository(
     /// <returns>A Result instance indicating success or failure.</returns>
     public async Task<Result<Comment>> UpdateAsync(int commentId, Comment comment)
     {
-        await using var context = await contextFactory.CreateDbContextAsync();
+        await using var context = await dbContextFactory.CreateDbContextAsync();
         var existing = await context.Comments.FirstOrDefaultAsync(e => e.Id == commentId);
         if (existing is null) return Result.Fail<Comment>(new NotFoundError());
