+++ FosterRoster/Features/Comments/ConfirmDeleteComment.razor
@@ -1,0 +1,23 @@
+@inject DialogService DialogService
+
+<RadzenStack>
+    <RadzenRow>
+        <RadzenColumn>
+            <RadzenText Typo="Typo.body1">
+                If you click Yes, you will delete all this comment.
+                <em>This action cannot be undone.</em>
+            </RadzenText>
+        </RadzenColumn>
+    </RadzenRow>
+    <RadzenStack Orientation="Orientation.Horizontal" JustifyContent="JustifyContent.End">
+        <RadzenButton ButtonStyle="ButtonStyle.Base" Click="@Cancel" Variant="Variant.Text">Cancel</RadzenButton>
+        <RadzenButton ButtonStyle="ButtonStyle.Danger" Click="@Confirm" Variant="Variant.Text">Yes</RadzenButton>
+    </RadzenStack>
+</RadzenStack>
+
+@code {
+    [Parameter] public string Name { get; set; } = string.Empty;
+
+    void Cancel() => DialogService.Close(false);
+    void Confirm() => DialogService.Close(true);
+}
