+++ FosterRoster/Features/Felines/FelineForm.razor
@@ -1,7 +1,7 @@
 @inject DialogService DialogService
 
-@using FosterRoster.Features.Sources
 @using FosterRoster.Features.Fosterers
+@using FosterRoster.Features.Sources
 
 @if (_model is null)
 {
