+++ FosterRoster/Features/Felines/FelinePrintView.razor
@@ -1,4 +1,6 @@
 ﻿@page "/felines/{FelineId:int}/print"
+@using Features.Comments
+@using Features.Thumbnails
 @using Features.Weights
 @layout Shared.Layout.PrintLayout
 @attribute [ExcludeFromInteractiveRouting]
