+++ FosterRoster/FosterRoster.csproj
@@ -32,4 +32,9 @@
         <PackageReference Include="SharpGrip.FluentValidation.AutoValidation.Mvc" Version="1.5.0"/>
     </ItemGroup>
 
+    <ItemGroup>
+      <AdditionalFiles Include="Features\Comments\CommentCard.razor" />
+      <AdditionalFiles Include="Features\Comments\CommentForm.razor" />
+    </ItemGroup>
+
 </Project>
