--- FosterRoster/Infrastructure/QueryResults.cs
@@ -1,3 +1,0 @@
-﻿namespace FosterRoster.Infrastructure;
-
-public sealed record QueryResults<TEntity>(List<TEntity> Items, int Count) where TEntity : class;
