+++ FosterRoster/Shared/Layout/AppNavMenu.razor
@@ -57,9 +57,7 @@
     private string? _currentUrl;
     private RadzenMenu _menu = null!;
 
-    [CascadingParameter] private HttpContext? HttpContext { get; set; } = null;
-
-    private bool IsInteractive => HttpContext?.AcceptsInteractiveRouting() ?? false;
+    [CascadingParameter] private HttpContext? HttpContext { get; set; }
 
     protected override void OnInitialized()
     {
